sql queries I used to generate two new tables
first table ('all_olympics') looks through the tables for each olympics from 2020 back to 1996 and creates a new table with one column, containing all the countries present at least once in the past 7 olympics.
second table ('oly_rank_and_data') went through all olympics back to 1996 and put the rank for each country in 'all_olympics' in to a new column, so each countries ranking for the past 7 olympics was avaialble in one table.
note: 1996 was chosen because this seems to be the first olympics where most of the modern nation states are present, before that there were many countrys (USSR and associated states) that no longer exist